Updated a large number of things including supporting back end change…

+++ src/app/component/review-container/review-container.component.ts
@@ -3,6 +3,7 @@ import { Movie } from 'src/app/model/movie';
 import { Review } from 'src/app/model/review';
 import { ReviewService } from 'src/app/service/review.service';
 import { Router } from '@angular/router';
+import { SessionService } from 'src/app/service/session.service';
 
 
 @Component({
@@ -17,9 +18,20 @@ export class ReviewContainerComponent {
   @Input() format: number; // Format 1: Has no user link, Format 2: has delete button and no open review, Format 3: has no user link and no open review
   uname: String;
 
-  constructor(private reviewService: ReviewService, private router: Router){}
+  constructor(private sessionService: SessionService, private reviewService: ReviewService, private router: Router){}
   onDeleteReview(revid) {
-    this.reviewService.deletereview(revid);
+    this.sessionService.checksession().subscribe((userFound: boolean) => {
+      if (userFound) {
+        console.log("User found, deleting review and redirecting to home.");
+        this.reviewService.deletereview(revid);
+        this.router.navigateByUrl('/');
+      } else {
+        localStorage.setItem('previousRoute', this.router.url);
+        console.log("User not found redirecting to login form.");
+        this.router.navigateByUrl('/loginform');
+      }
+    });
+    
   }
   openReview(revid: string): void {
     this.router.navigate(['/review'], { queryParams: { revid: revid } });
